requirements.txt
Flask
libvirt-python
flask-login